+++ R/report.R
@@ -40,35 +40,29 @@
 
 report <- function(object, file = 'report.html', format = 'html_document'){
   
-  if (!class(object) == "jfaEvaluation")
+  if (!class(object) == 'jfaEvaluation')
     stop("Object must be of class 'jfaEvaluation'.")
   
-  if (!requireNamespace("rmarkdown", quietly = TRUE)) {
-    stop("Package \"rmarkdown\" needed for this function to work. Please install it.",
-         call. = FALSE)
-  }
+  if (!requireNamespace('rmarkdown', quietly = TRUE))
+    stop('Package \"rmarkdown\" needed for this function to work. Please install it.', call. = FALSE)
   
-  if (!requireNamespace("knitr", quietly = TRUE)) {
-    stop("Package \"knitr\" needed for this function to work. Please install it.",
-         call. = FALSE)
-  }
+  if (!requireNamespace('knitr', quietly = TRUE))
+    stop('Package \"knitr\" needed for this function to work. Please install it.', call. = FALSE)
   
-  if (!requireNamespace("kableExtra", quietly = TRUE)) {
-    stop("Package \"kableExtra\" needed for this function to work. Please install it.",
-         call. = FALSE)
-  }
+  if (!requireNamespace('kableExtra', quietly = TRUE))
+    stop('Package \"kableExtra\" needed for this function to work. Please install it.', call. = FALSE)
   
-  #Determine the template
-  theFile <- system.file("rmd/report.Rmd", package = "jfa")
+  # Determine the template
+  theFile <- system.file('rmd/report.Rmd', package = 'jfa')
   
-  #Process the Arguments
+  # Process the function arguments
   args               <- list()
   args$input         <- theFile
   args$output_dir    <- getwd()
   args$output_format <- format
   args$output_file   <- file
   
-  #Run the render
+  # Start the renderer via rmarkdown
   outputFileName <- do.call(.getfun('rmarkdown::render'), args = args)
   invisible(outputFileName)
 }